Move readers for test report format out of selenium module
