+++ bzt/modules/selenium.py
@@ -22,10 +22,10 @@ from bzt import TaurusConfigError, TaurusInternalException
 from urwid import Text, Pile
 
 from bzt.engine import Scenario, FileLister, SubprocessedExecutor
-from bzt.modules.aggregator import ConsolidatingAggregator, ResultsReader
+from bzt.modules.aggregator import ConsolidatingAggregator
 from bzt.modules.console import WidgetProvider, PrioritizedWidget
-from bzt.modules.functional import FunctionalResultsReader, FunctionalAggregator, FunctionalSample
-from bzt.six import string_types
+from bzt.modules.functional import FunctionalAggregator, FuncSamplesReader, LoadSamplesReader
+from bzt.modules.python import NoseTester
 from bzt.utils import is_windows, BetterDict, get_full_path, get_files_recursive
 
 try:
@@ -174,11 +174,10 @@ class SeleniumExecutor(AbstractSeleniumExecutor, WidgetProvider, FileLister):
         self.scenario = self.get_scenario()
         self.script = self.get_script_path()
 
-        script_type = self.detect_script_type()
         self.report_file = self.engine.create_artifact("selenium_tests_report", ".ldjson")
         self.runner = self._create_runner(self.report_file)
         self.runner.prepare()
-        if script_type == "nose":
+        if isinstance(self.runner, NoseTester):
             self.script = self.runner._script
         if self.register_reader:
             self.reader = self._register_reader(self.report_file)
@@ -328,145 +327,3 @@ class SeleniumWidget(Pile, PrioritizedWidget):
             self.summary_stats.set_text('In progress...')
 
         self._invalidate()
-
-
-class LDJSONReader(object):
-    def __init__(self, filename, parent_log):
-        self.log = parent_log.getChild(self.__class__.__name__)
-        self.filename = filename
-        self.fds = None
-        self.partial_buffer = ""
-        self.offset = 0
-
-    def read(self, last_pass=False):
-        if not self.fds and not self.__open_fds():
-            self.log.debug("No data to start reading yet")
-            return
-
-        self.fds.seek(self.offset)
-        if last_pass:
-            lines = self.fds.readlines()  # unlimited
-        else:
-            lines = self.fds.readlines(1024 * 1024)
-        self.offset = self.fds.tell()
-
-        for line in lines:
-            if not line.endswith("\n"):
-                self.partial_buffer += line
-                continue
-            line = "%s%s" % (self.partial_buffer, line)
-            self.partial_buffer = ""
-            yield json.loads(line)
-
-    def __open_fds(self):
-        if not os.path.isfile(self.filename):
-            return False
-        fsize = os.path.getsize(self.filename)
-        if not fsize:
-            return False
-        self.fds = open(self.filename, 'rt', buffering=1)
-        return True
-
-    def __del__(self):
-        if self.fds is not None:
-            self.fds.close()
-
-
-class SeleniumReportReader(object):
-    REPORT_ITEM_KEYS = ["test_case", "test_suite", "status", "start_time", "duration",
-                        "error_msg", "error_trace", "extras"]
-    TEST_STATUSES = ("PASSED", "FAILED", "BROKEN", "SKIPPED")
-    FAILING_TESTS_STATUSES = ("FAILED", "BROKEN")
-
-    def __init__(self, filename, parent_logger, translation_table=None):
-        super(SeleniumReportReader, self).__init__()
-        self.log = parent_logger.getChild(self.__class__.__name__)
-        self.json_reader = LDJSONReader(filename, self.log)
-        self.translation_table = translation_table or {}
-
-    def process_label(self, label):
-        if label in self.translation_table:
-            return self.translation_table[label]
-
-        if isinstance(label, string_types):
-            if label.startswith('test_') and label[5:10].isdigit():
-                return label[11:]
-
-        return label
-
-    def read(self, last_pass=False):
-        for row in self.json_reader.read(last_pass):
-            for key in self.REPORT_ITEM_KEYS:
-                if key not in row:
-                    self.log.debug("Unexpected test record: %s", row)
-                    self.log.warning("Test record doesn't conform to schema, skipping, %s", key)
-                    continue
-
-            row["test_case"] = self.process_label(row["test_case"])
-            yield row
-
-
-class LoadSamplesReader(ResultsReader):
-    STATUS_TO_CODE = {
-        "PASSED": "200",
-        "SKIPPED": "300",
-        "FAILED": "400",
-        "BROKEN": "500",
-    }
-
-    def __init__(self, filename, parent_logger, translation_table):
-        super(LoadSamplesReader, self).__init__()
-        self.report_reader = SeleniumReportReader(filename, parent_logger, translation_table)
-        self.read_records = 0
-
-    def extract_sample(self, item):
-        tstmp = int(item["start_time"])
-        label = item["test_case"]
-        concur = 1
-        rtm = item["duration"]
-        cnn = 0
-        ltc = 0
-        rcd = self.STATUS_TO_CODE.get(item["status"], "UNKNOWN")
-        error = item["error_msg"] if item["status"] in SeleniumReportReader.FAILING_TESTS_STATUSES else None
-        trname = ""
-        byte_count = None
-        return tstmp, label, concur, rtm, cnn, ltc, rcd, error, trname, byte_count
-
-    def _read(self, last_pass=False):
-        for row in self.report_reader.read(last_pass):
-            self.read_records += 1
-            sample = self.extract_sample(row)
-            yield sample
-
-
-class FuncSamplesReader(FunctionalResultsReader):
-    FIELDS_EXTRACTED_TO_ARTIFACTS = ["requestBody", "responseBody", "requestCookiesRaw"]
-
-    def __init__(self, filename, engine, parent_logger, translation_table):
-        self.report_reader = SeleniumReportReader(filename, parent_logger, translation_table)
-        self.engine = engine
-        self.read_records = 0
-
-    def _write_sample_data_to_artifacts(self, sample_extras):
-        if not sample_extras:
-            return
-        for file_field in self.FIELDS_EXTRACTED_TO_ARTIFACTS:
-            if file_field not in sample_extras:
-                continue
-            contents = sample_extras.pop(file_field)
-            if contents:
-                filename = "sample-%s" % file_field
-                artifact = self.engine.create_artifact(filename, ".bin")
-                with open(artifact, 'wb') as fds:
-                    fds.write(contents.encode('utf-8'))
-                sample_extras[file_field] = artifact
-
-    def read(self, last_pass=False):
-        for row in self.report_reader.read(last_pass):
-            self.read_records += 1
-            sample = FunctionalSample(test_case=row["test_case"], test_suite=row["test_suite"],
-                                      status=row["status"], start_time=row["start_time"], duration=row["duration"],
-                                      error_msg=row["error_msg"], error_trace=row["error_trace"],
-                                      extras=row.get("extras", {}))
-            self._write_sample_data_to_artifacts(sample.extras)
-            yield sample
