+++ bzt/utils.py
@@ -1085,3 +1085,45 @@ def humanize_bytes(byteval):
     # (.4g results in rounded numbers for exact matches and max 3 decimals,
     # should never resort to exponent values)
     return '{:.4g}{}'.format(byteval / (1 << (order * 10)), _suffixes[order])
+
+
+class LDJSONReader(object):
+    def __init__(self, filename, parent_log):
+        self.log = parent_log.getChild(self.__class__.__name__)
+        self.filename = filename
+        self.fds = None
+        self.partial_buffer = ""
+        self.offset = 0
+
+    def read(self, last_pass=False):
+        if not self.fds and not self.__open_fds():
+            self.log.debug("No data to start reading yet")
+            return
+
+        self.fds.seek(self.offset)
+        if last_pass:
+            lines = self.fds.readlines()  # unlimited
+        else:
+            lines = self.fds.readlines(1024 * 1024)
+        self.offset = self.fds.tell()
+
+        for line in lines:
+            if not line.endswith("\n"):
+                self.partial_buffer += line
+                continue
+            line = "%s%s" % (self.partial_buffer, line)
+            self.partial_buffer = ""
+            yield json.loads(line)
+
+    def __open_fds(self):
+        if not os.path.isfile(self.filename):
+            return False
+        fsize = os.path.getsize(self.filename)
+        if not fsize:
+            return False
+        self.fds = open(self.filename, 'rt', buffering=1)
+        return True
+
+    def __del__(self):
+        if self.fds is not None:
+            self.fds.close()
